+++ src/main/java/org/marble/util/Connectors.java
@@ -1,22 +1,22 @@
 package org.marble.util;
 
-import javax.vecmath.Matrix4f;
+import javax.vecmath.Matrix4d;
 
 import com.google.common.collect.ImmutableMap;
 
 import org.marble.entity.Connector;
 
 public final class Connectors {
-    public static ImmutableMap<String, Connector> fromBox(final float width,
-            final float height, final float depth) {
+    public static ImmutableMap<String, Connector> fromBox(final double width,
+            final double height, final double depth) {
 
         final int xcount = (int) width, ycount = (int) height, zcount =
                 (int) depth;
-        final float xborder = (width - xcount + 1) / 2, yborder =
+        final double xborder = (width - xcount + 1) / 2, yborder =
                 (height - ycount + 1) / 2, zborder = (depth - zcount + 1) / 2;
-        final float xhalf = width / 2, yhalf = height / 2, zhalf = depth / 2;
-        final float pi = (float) Math.PI;
-        final float pihalf = pi / 2;
+        final double xhalf = width / 2, yhalf = height / 2, zhalf = depth / 2;
+        final double pi = Math.PI;
+        final double pihalf = pi / 2;
         final ImmutableMap.Builder<String, Connector> connectorBuilder =
                 ImmutableMap.builder();
 
@@ -28,11 +28,11 @@ public final class Connectors {
         connectorBuilder.put("south_middle",
                 offsetBy(0, -yhalf, 0, -pihalf, 0, 0));
         for (int x = 0; x < xcount; x++) {
-            final float xcoord = xborder + x - xhalf;
+            final double xcoord = xborder + x - xhalf;
 
             // Top/Bottom
             for (int y = 0; y < ycount; y++) {
-                final float ycoord = yborder + y - yhalf;
+                final double ycoord = yborder + y - yhalf;
                 connectorBuilder.put(makeAnchorName("top", x, y),
                         offsetBy(xcoord, ycoord, zhalf, 0, pihalf, 0));
                 connectorBuilder.put(makeAnchorName("bottom", x, y),
@@ -41,7 +41,7 @@ public final class Connectors {
 
             // North/South
             for (int z = 0; z < zcount; z++) {
-                final float zcoord = zborder + z - zhalf;
+                final double zcoord = zborder + z - zhalf;
                 connectorBuilder.put(makeAnchorName("north", x, z),
                         offsetBy(xcoord, yhalf, zcoord, pihalf, 0, 0));
                 connectorBuilder.put(makeAnchorName("south", x, z),
@@ -53,9 +53,9 @@ public final class Connectors {
         connectorBuilder.put("east_middle", offsetBy(xhalf, 0, 0, 0, 0, 0));
         connectorBuilder.put("west_middle", offsetBy(-xhalf, 0, 0, pi, 0, 0));
         for (int y = 0; y < ycount; y++) {
-            final float ycoord = yborder + y - yhalf;
+            final double ycoord = yborder + y - yhalf;
             for (int z = 0; z < zcount; z++) {
-                final float zcoord = zborder + z - yhalf;
+                final double zcoord = zborder + z - yhalf;
                 connectorBuilder.put(makeAnchorName("east", y, z),
                         offsetBy(xhalf, ycoord, zcoord, 0, 0, 0));
                 connectorBuilder.put(makeAnchorName("west", y, z),
@@ -71,11 +71,11 @@ public final class Connectors {
         return base + "_" + coord1 + "_" + coord2;
     }
 
-    public static Connector
-            offsetBy(final float x, final float y, final float z,
-                    final float yaw, final float pitch, final float roll) {
-        final Matrix4f rotation = new Matrix4f();
-        final Matrix4f transform = new Matrix4f();
+    public static Connector offsetBy(final double x, final double y,
+            final double z, final double yaw, final double pitch,
+            final double roll) {
+        final Matrix4d rotation = new Matrix4d();
+        final Matrix4d transform = new Matrix4d();
 
         // TODO make more efficient; we don't need a temp matrix here.
         transform.setIdentity();
@@ -85,28 +85,6 @@ public final class Connectors {
         transform.mul(rotation);
         rotation.rotY(roll);
         transform.mul(rotation);
-
-        /**
-         * XXX This calculation is buggy; yaw=2pi does for example collapse the
-         * matrix {@code
-         * final double sinPitch = sin(pitch);
-         * final double cosPitch = cos(pitch);
-         * final double sinRoll = sin(roll);
-         * final double cosRoll = cos(roll);
-         * final double sinYaw = sin(yaw);
-         * final double cosYaw = cos(yaw);
-         * transform.setElement(0, 0, (float) (cosPitch * cosYaw));
-         * transform.setElement(0, 1, (float) (-sinYaw * cosPitch));
-         * transform.setElement(0, 2, (float) sinPitch); transform.setElement(1,
-         * 0, (float) (cosYaw * sinPitch * sinRoll + sinYaw * cosRoll));
-         * transform.setElement(1, 1, (float) (-sinYaw * sinPitch * sinRoll +
-         * cosYaw * cosRoll)); transform.setElement(1, 2, (float) (-cosPitch *
-         * sinRoll)); transform.setElement(2, 0, (float) (-cosYaw * sinPitch *
-         * cosRoll + sinYaw * sinRoll)); transform.setElement(2, 1, (float)
-         * (sinYaw * sinPitch * cosRoll + cosYaw * sinRoll));
-         * transform.setElement(2, 2, (float) (cosPitch * cosRoll));
-         * }
-         */
 
         // Translation column
         transform.setElement(0, 3, x);
